refactor: moved the Fetch command
