+++ src/commands/mod.rs
@@ -7,7 +7,8 @@ mod run;
 
 // Re-export only the functions that execute commands
 pub use data::data_description::DataStatus;
-pub use data::{clean_data, fetch_remote_data, package_data_to_archive};
+pub use data::fetch::FetchCommand;
+pub use data::{clean_data, package_data_to_archive};
 
 pub use new::create_kerblam_project;
 pub use other::ignore;
